Use suggestion type correctly in UI
Although some code treated the suggestion type as a bitstring of flags,
not all of it did. I have modified all logic accessing the types to
treat them as they are defined in the Python back end.

One notable change is that all received suggestion types from
`has_suggestions` messages are now OR'ed together. Before this commit,
the last seen value would be used. This adjustment allows a user to be
advised by multiple advisors at once, which likely was not considered
when this feature was first implemented.