Migrate from `setup.py` to `pyproject.toml`
`pyproject.toml` is the latest standard, and using it will make it
easier to change requirements, which a future commit intends to do. I
took this opportunity to slightly modernize some of the package
metadata, but those changes were small.

I based the new `pyproject.toml` on the one in `chiron-utils`.[^1]

[^1]: https://github.com/ALLAN-DIP/chiron-utils/blob/e095a92ead136dde8a63e2ed76c4e11712e46877/pyproject.toml