Improve `.dockerignore`
I discovered that I wrote the file without using the proper
`.dockerignore` syntax,[^1] so Docker was including most of the files I
was trying to exclude! I added leading `**/`s where relevant to fix the
problem.

I also copied some lines from `chiron-utils`[^2] to exclude some more
files that are also in `.gitignore`.

This commit reduces the size of the image (as reported by `docker
images` on macOS) from 102 MB to 37.9 MB.

[^1]: https://docs.docker.com/build/concepts/context/#syntax
[^2]: https://github.com/ALLAN-DIP/chiron-utils/blob/f9f206953ff3109dc0b03310dc86e7b5c0962592/.dockerignore