Read order distribution from new message type
This replaces the call to the API that no longer exists.

We need to pass the province's power directly into
`onChangeOrderDistribution()` because the message does not include that
information. I didn't realize it was needed when I designed the new
message type's format, and it's public information that should be read
directly from the game state, anyway.