Merge roughly duplicated conditionals
If the move advice component is enabled, but there is currently no move
advice to display, then the component will not be displayed. This
situation is expected to happen in some cases, such as when an advisor
takes a minute or two to generate move advice. We should just display
the component in all situations where move advice is expected to prevent
the component from appearing and disappearing as advice arrives or is
dismissed.