Do not read username from `localStorage`
This was causing a log panel to be displayed for a normal player if they
signed into an account named `admin` in another tab in the same browser.
We already read the username from session information elsewhere on the
page, so we should just read from it here as well.