Factor out logging code from `__init__.py`
I discovered some `Logging error` messages that showed up when running
`pytest`, and I decided to look into them in case they hid an actual
bug.

According to the Python documentation,[^1][^2] libraries shouldn't
modify logging configuration because it removes the ability for
applications to properly configure logging. `diplomacy` now only
initializes logging when run as an application, not imported as a
library.

[^1]: https://docs.python.org/3.13/howto/logging.html#configuring-logging-for-a-library
[^2]: https://docs.python.org/3.13/howto/logging-cookbook.html#adding-handlers-other-than-nullhandler-to-a-logger-in-a-library