Disable message sending buttons if orders are unset
The text box was already made uneditable, but text entered from a
previous turn could still be sent. The buttons are now disabled to
prevent sending messages when we don't want them to be sent.

An alternative option was to delete any text in the message box at the
end of the turn, but I decided not to do that because the user might
want to be able to send their typed-out message the next turn.