Disable message advice buttons if orders are unset
The "add to textbox" button allowed one to bypass the check that
prevented sending messages without orders. The buttons should now be
hidden when orders are unset, but to prevent any regressions, the
buttons are now also disabled.

The "✕" button is now similarly disabled, but that is only done for
consistency with the "add to textbox" button.