Fix order removal by human player not being sent to advisor
This bug expressed itself by not allowing human players to receive
partial order advice. The web UI does not show any advice until the
player has submitted orders, so the player only saw full order advice.
The only way to receive partial order advice was to clear all orders and
add some again, but that is unintuitive and shouldn't be required when
individual orders can be removed.

The problem was that `Game.set_orders()` takes in a list of orders and
only removes an existing order if a replacement order for the same unit
is provided. My initial idea was to treat the list of orders as complete
instead of partial, but that breaks so much existing code that I decided
to find a better solution.

`Game.set_orders()` is called in three functions relevant to order
submission:

- The `on_set_orders()` request manager sets new orders on the server,
  calling `Power.clear_orders()` first to ensure that orders removed by
  the client are removed everywhere.
- The `on_set_orders()` response manager sets new orders on the client
  that made the request to remove orders.
- The `on_power_orders_update()` notification manager sets new orders on
  other clients for the same power, as well as all observing clients.
  The notification is triggered by the `on_set_orders()` request manager
  on the server.

The solution here ended up being surprisingly simple: call
`Power.clear_orders()` in `on_power_orders_update()`, which is the same
thing the `on_set_orders()` request manager does. This ends up only
affecting the advisor, which is what we want.