Fix `SuggestionType.to_parsable()` for Python 3.11
I created the function back when we were running `diplomacy` with Python
3.7 and only used the function in Cicero (which also uses Python 3.7),
so I never realized that the function didn't work properly in Python
3.11. It took until I tried using the function in `chiron-utils`, which
uses Python 3.11, for me to notice!

The problem is that `IntFlag.__str__()` was changed in Python 3.11,
which messed up my logic and caused `to_parsable()` to output the empty
string instead of the value it was supposed to. I therefore rewrote
`to_parsable()` to work the same for both Python 3.7 and 3.11. The new
implementation changed the order of members in the generated string, so
I changed the docstrings accordingly.

To prevent missing any more bugs, I wrote some tests. They helped me
find multiple bugs in my implementation. Now that all tests pass, I'm
_pretty_ confident that my implementation is correct!